Add REST API with request plumbing and a data loader
Serve /api/v1 over Go's own mux, which routes on method and path patterns
since 1.22, so a third-party router would be a dependency taken on for
syntax alone. Adds fail-fast configuration, request and correlation id
propagation, structured access logging, panic recovery, an explicit CORS
allowlist, RFC 7807 errors and keyset pagination.

Endpoints cover athletes, their analytics, sessions, session timelines,
cross-session pitch search, predictions, model cards and anomaly
acknowledgement, plus liveness and readiness.

Decisions a reviewer would push on:

- GET /pitches requires pitcher_id. Without it the query is a sequential
  scan over millions of rows, and refusing to express the expensive query
  beats optimizing it after someone finds it in production.

- Cursors always carry an id alongside the sort key. Replay can emit
  several pitches with the same timestamp, and paginating on time alone
  silently drops all but one at a page boundary.

- Session close is an action endpoint, not a PATCH on status. It is a state
  transition with side effects, and a second call returns 409 rather than a
  silent success, because "already done" and "never existed" are different
  answers.

- Empty aggregates return null. SQL gives 0 for avg() over no rows, which
  would claim a pitcher throws 0 mph.

- Liveness checks nothing. A liveness probe that touches the database turns
  one slow query into a restart storm. Readiness does the checking, and
  reports Redis as degraded while still returning 200, since a cache outage
  must not take the API out of rotation.

Three type inference problems surfaced while wiring the generated queries.
A division guard came back as an integer, aggregates were inferred
non-nullable and would have panicked scanning the NULL an empty group
produces, and min/max over a subquery came back as interface{}. The first
two are fixed with explicit casts and COALESCE; the third is converted in
Go, because casting it in SQL only moves the problem to scan time on
exactly the rows that are null.

Integration tests now run with -p 1. They share one database and each
truncates it, so parallel packages were deleting each other's fixtures and
concurrent TRUNCATEs deadlocked.

Also adds a loader that puts real pitches in PostgreSQL so the API has
something to serve before Kafka exists. It mirrors what the stream
processor will do -- deterministic session ids, the same natural key, the
same normalization code, and the outcome written only to actual_outcome --
so the two cannot drift before the simulator replaces it.

+++ internal/db/queries/athletes.sql
@@ -45,8 +45,12 @@ ORDER BY a.full_name, a.id
 LIMIT $1;
 
 -- name: GetAthleteTotals :one
+-- first_pitch_at and last_pitch_at are genuinely NULL for an athlete with no
+-- pitches. They are left uncast: an explicit ::timestamptz makes sqlc emit a
+-- non-pointer time.Time, which then fails at scan time on exactly the rows
+-- that need handling. The interface{} sqlc infers is converted in the handler.
 SELECT
-    (SELECT count(*) FROM sessions s WHERE s.pitcher_athlete_id = $1) AS session_count,
-    (SELECT count(*) FROM pitches p  WHERE p.pitcher_athlete_id = $1) AS pitch_count,
+    (SELECT count(*) FROM sessions s WHERE s.pitcher_athlete_id = $1)::bigint AS session_count,
+    (SELECT count(*) FROM pitches p  WHERE p.pitcher_athlete_id = $1)::bigint AS pitch_count,
     (SELECT min(p.thrown_at) FROM pitches p WHERE p.pitcher_athlete_id = $1) AS first_pitch_at,
     (SELECT max(p.thrown_at) FROM pitches p WHERE p.pitcher_athlete_id = $1) AS last_pitch_at;
